Module Ez_logs

Logs reporting

val reporter : ?max_level:Logs.level -> ?time_precision:int -> Stdlib.Format.formatter -> Logs.reporter * Logs.level option Stdlib.ref

reporter ?max_level ?time_precision ppf returns both a properly setup Logs reporter for formatted output on ppf, and a reference to the level below which the reporter does not report log messages.

The returned referenced level !max_level is used to filter out messages: if it is None, nothing is printed; if it is Some l, then every message whose level is above l is printed.

time_precision is used to enable and control a timestamp prefix: if it is 0 (the default) or negative, then no timesamp is prepended. Otherwise if it is p, then each message is prepended with the number of seconds elapsed since the tool process started, measured with precision p and put in between square brackets.

val combine_reporters : (Logs.reporter as 'r) -> 'r -> 'r

Combines two reporters so log messages can be broadcast to multiple reporters.

val persist_in : oc:Stdlib.out_channel -> ?time_precision:int -> ?style_renderer:Fmt.style_renderer -> ?utf_8:bool -> ?right_margin:int -> Logs.reporter -> unit -> unit

persist_in ~oc ~time_precision reporter starts persisting logs that are output via reporter into output stream oc. Returns a function that should be called to stop reporting on oc.

val stdout_level_ref : Logs.level option Stdlib.ref

Reference to the log level associated with the reporter on standard output. In effect only once init_stdout_reporter has been called.

val init_stdout_reporter : ?max_level:Logs.level -> ?time_precision:int -> unit -> unit

init_stdout_reporter ~max_level ~time_precision () initializes a reporter on standard output. The initial level associated is given via max_level (None by default, meaning that nothing is reported on standard output), and can be updated via stdout_level_ref once the call returns.

module Types : module type of Types

Module signatures for unified logging interfaces.

module type T = Types.T

Alias for the types of combined log emitter interfaces

module From_src (_ : Types.SRC) : T

Builds a combined emitter from a given source

val from_src : ?styling:Fmt.style list -> Logs.src -> (module T)

Functional version of From_src

val subproc : ?format_log_name:(string -> string) -> string -> (module T)

subproc ?format_log_name procname returns a module suitable for logging outputs of sub-processes named procname. The name used for log entries is formatted using format_log_name (default is Format.asprintf "(%s)").

val anonymous_subproc : string -> (module T)